requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser automation (for SSO login)
selenium>=4.15.0
webdriver-manager>=4.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Authentication & Security
cryptography>=41.0.0

# Output generation
openpyxl>=3.1.0
tabulate>=0.9.0

# Progress & Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
